+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 5.9
 - Hide Permissions for "No Access" profile in General_Info.inc.php
 - Move Create Student/User Account and New Administrator notifications to functions in SendNotification.fnc.php
 - Send Account Activation email notification to Student in SendNotification.fnc.php & Student.php
+- Send Account Activation email notification to User in SendNotification.fnc.php & User.php
 
 Changes in 5.8
 --------------